4-23 Workday
-HOT/DOT system mostly complete (DOTs currently stack not sure how to disable this)
-first draft documentation done

+++ Assets/Scripts/Adam Scripts/Interfaces/IGunBehavior.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IGunBehavior
 {
-    void FireGun(Transform shootPoint, int damage, float range);
+    void FireGun(Transform shootPoint, float damage, float range);
 }
