+++ Assets/Scripts/Adam Scripts/WeaponScripts/Range Weapons/ContinuousGun.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class ContinuousGun : GunBase
 {
     //Constructor instantiates new ContinuousBehvaior then initializes ammoMax and ammoCount based on params
-    public ContinuousGun(GameObject fireVisualPrefab, int initialAmmoMax, int initialAmmoCount, int damage)
+    public ContinuousGun(GameObject fireVisualPrefab, int initialAmmoMax, int initialAmmoCount, float damage)
     {
         weaponBehavior = new ContinuousBehavior { fireVisualPrefab = fireVisualPrefab};
         ammoMax = initialAmmoMax;
